==> 1_data_analyst_dot_com.ipynb <==
import re
import time
import requests
import numpy as np

from bs4 import BeautifulSoup
from collections import Counter
from selenium import webdriver

from wordcloud import WordCloud, STOPWORDS
import matplotlib.pyplot as plt
import pandas as pd

from selenium.webdriver.common.by import By


url = 'https://www.dataanalyst.com/?ea8d9d2b_page='

urls = []
for i in range(1,8):
    urls.append(url+str(i))
    
print(urls[0])
range(len(urls))

## launch headless Edge instance
options = webdriver.EdgeOptions()
options.add_argument('--headless')


## initiate variables for loop
df = pd.DataFrame(columns = ['job_title', 'job_url','job_country', 'job_company', 'job_location', 'job_workplace', 'job_industry', 'job_experience', 'job_salary', 'job_published', 'job_roles', 'job_requirements'])
i = 0
for ii in range(len(urls)): 
    
    print("Starting")
    print(ii)
    
    driver = webdriver.Edge(options = options)
    driver.get(urls[ii])

    ## grab html from page
    page = driver.page_source.encode('utf-8')
    soup = BeautifulSoup(page, 'html.parser')

    job_rows = soup.find_all('a', class_ = 'card job cms-jobs-counter w-inline-block') 
    print("Number of jobs")
    print(len(job_rows))

    ## loop through rows, grabbing unique job posting url from each  
    for iii in range(len(job_rows)): # 
        
        ## grab url for unique job post
        job_link = re.search(r'(?<=href=\").*(?=\"><div class=\"html-embed-5 w-embed\")', str(job_rows[iii])).group()
        print(job_link)
        j_url = 'https://www.dataanalyst.com/'
        job_url = str(j_url + 
                      re.search(r'(?<=href=\").*(?=\"><div class=\"html-embed-5 w-embed\")', 
                                str(job_rows[iii])).group()
                     )

        # Lets try getting everything in a list
        a = re.findall(r'(?<=\>)[0-9a-zA-z\s\,\.\-]+(?=\<)',str(job_rows[iii]))

        # title 0, company 1
        job_title = a[0]
        job_company = a[1]
        # Country
        job_country = a[a.index("Country") + 1]
        # Location
        job_location = a[a.index("Location") + 1]
        # Workplace
        job_workplace = a[a.index("Workplace") + 1]
        # Industry
        job_industry = a[a.index("Industry") + 1]
        # published
        job_published = a[a.index("Published") + 1]

        # Still need experience, salary and currency
        # Experience
        a = re.findall(r'(?<=\>)[0-9]+.*[0-9]+.*(?=years)',str(job_rows[iii]))
        job_experience = a
        # Salary
        try:
            a = re.search(r'(?<= class=\"card-job-category-title-wrapper\"\>\<div\>Salary\<\/div\>\<\/div\>\<div class=\"card-job-category-text\">).*(?=\<\/div\>\<\/div\>\<div class=\"card-job-category-wrapper-industry\")', str(job_rows[iii])).group()
            job_salary = a
        except:
            job_salary = 0

        ## capture html from post page
        job_page = requests.get(job_url)
        job_soup = BeautifulSoup(job_page.content, 'html.parser')

        # grab job roles
        job_roles = job_soup.find('div', class_ = 'rich-text-block-3 w-richtext')
        job_roles = re.sub(r'\<.*\>', '', job_roles.prettify())

        # grab job responsibilities
        job_requirements = job_soup.find('div', class_ = 'rich-text-block-4 w-richtext')
        job_requirements = re.sub(r'\<.*\>', '', job_requirements.prettify())

        ## move all variables to dataframe
        df.loc[i, 'job_title'] = job_title
        df.loc[i,'job_url'] = job_url
        df.loc[i,'job_company'] = job_company
        df.loc[i,'job_country'] = job_country
        df.loc[i,'job_location'] = job_location
        df.loc[i,'job_workplace'] = job_workplace
        df.loc[i,'job_industry'] = job_industry
        df.loc[i,'job_experience'] = job_experience
        df.loc[i,'job_salary'] = job_salary
        df.loc[i,'job_published'] = job_published
        df.loc[i,'job_roles'] = job_roles
        df.loc[i,'job_requirements'] = job_requirements

        time.sleep(1)
        
        i = i + 1
        
    df2 = df
    df2.job_salary = df2.job_salary.str.replace('[^0-9]','-')
    df2.job_roles = df2.job_roles.str.replace('[^a-zA-Z0-9#&\.\+\-]',' ')
    df2.job_requirements = df2.job_requirements.str.replace('[^a-zA-Z0-9#&\.\+\-]',' ')
    df2.to_csv("data_analyst_dot_com.csv")
        
    time.sleep(2)
