==> 1_nyc_jobs.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

## Read Data and Change File Type

- 3/6/2023: Importing the data from [NYC Open Data](https://data.cityofnewyork.us/City-Government/NYC-Jobs/pda4-rgn4/) as a CSV file.
- Transform file from CSV to parquet to compress file size

df_parquet = pd.read_parquet(r'data/nyc_jobs.parquet')

df_parquet.head(3)

###

df_parquet.columns

cols_to_keep = [1,4,8,9,10,13,14,15,16]

df_jobs = df_parquet.iloc[:, cols_to_keep]
df_jobs.head(3)
