==> 1_usa_jobs_data_analyst.ipynb <==
import re
import time
import requests
import numpy as np

from bs4 import BeautifulSoup
from collections import Counter
from selenium import webdriver

from wordcloud import WordCloud, STOPWORDS
import matplotlib.pyplot as plt
import pandas as pd

from selenium.webdriver.common.by import By


url = 'https://www.usajobs.gov/Search/Results?jt=Data%20Analyst'

## launch headless Edge instance
options = webdriver.ChromeOptions()
options.add_argument('--headless')

driver = webdriver.Chrome(options = options)
driver.get(url)

## grab html from page
page = driver.page_source.encode('utf-8')
soup = BeautifulSoup(page, 'html.parser')

## find total number of jobs found
job_count_ele = soup.find(id = 'page-info')
print(job_count_ele.prettify())


job_count = int(re.search(r'[0-9]+(?= jobs)', str(job_count_ele)).group())
print(job_count)

## initiate variables for loop
df = pd.DataFrame(columns = ['job_title', 'job_url','job_summary', 'job_salary', 'job_location', 'job_quals', 'job_duties'])
words = []
counter = 0
start = time.time()
i = 0
for ii in range(10): 
    
    print("Starting")
    print(ii)
    page = driver.page_source.encode('utf-8')
    soup = BeautifulSoup(page, 'html.parser')
    job_rows = soup.find_all('a', class_ = 'usajobs-search-result--core__title') 
    print("Number of jobs")
    print(len(job_rows))

    ## loop through rows, grabbing unique job posting url from each  
    for iii in range(len(job_rows)): # 
        
        ## grab url for unique job post
        job_link = re.search(r'(?<=href=\").*(?=\" id)', str(job_rows[iii])).group()
        print(job_link)
        j_url = 'https://www.usajobs.gov/'
        job_url = str(j_url + 
                      re.search(r'(?<=href=\").*(?=\" id)', 
                                str(job_rows[iii])).group()
                     )

        ## capture html from post page
        job_page = requests.get(job_url)
        job_soup = BeautifulSoup(job_page.content, 'html.parser')
        
        # grab job title
        job_title = job_soup.find('h1', class_ = 'usajobs-joa-banner__title')
        job_title = re.sub(r'\<.*\>', '', job_title.prettify())

        # grab job summary
        job_summary = job_soup.find('div', id = 'summary')
        job_summary = re.sub(r'\<.*\>', '', job_summary.prettify())
        
        # grab job description
        job_salary = job_soup.find('p', class_ = 'usajobs-joa-summary__salary salary-text-normal')
        job_salary = re.sub(r'\<.*\>', '', job_salary.prettify())
        
        # grab job location
        job_location = job_soup.find('div', class_ = "usajobs-joa-locations__body")
        job_location = re.sub(r'\<.*\>', '', job_location.prettify())
        
        # grab job qualifications
        job_quals = job_soup.find('div', id = "qualifications")
        job_quals = re.sub(r'\<.*\>', '', job_quals.prettify())
        
        job_duties = job_soup.find('div', id = "duties")
        job_duties = re.sub(r'\<.*\>', '', job_duties.prettify())
        
        
        ## move all variables to dataframe
        df.loc[i, 'job_title'] = job_title
        df.loc[i,'job_url'] = job_url
        df.loc[i,'job_summary'] = job_summary
        df.loc[i,'job_salary'] = job_salary
        df.loc[i,'job_location'] = job_location
        df.loc[i,'job_quals'] = job_quals
        df.loc[i,'job_duties'] = job_duties
        
        time.sleep(4)
        
        i = i + 1
        
    df2 = df
    df2.job_title = df2.job_title.str.replace('[^a-zA-Z0-9#&\.\+\-]', ' ')
    df2.job_summary = df2.job_summary.str.replace('[^a-zA-Z0-9#&\.\+\-]', ' ')
    df2.job_location = df2.job_location.str.replace('[^a-zA-Z0-9#&\.\+\-]', ' ')
    df2.job_quals = df2.job_quals.str.replace('[^a-zA-Z0-9#&\.\+\-]', ' ')
    df2.job_salary = df2.job_salary.str.replace('[^0-9]','-')
    df2.job_duties = df2.job_duties.str.replace('[^a-zA-Z0-9#&\.\+\-]',' ')
    df2.to_csv("data_analyst.csv")
        
    button = driver.find_element(By.CLASS_NAME, 'usajobs-search-pagination__next-page')
    driver.execute_script('arguments[0].click()', button)
    time.sleep(2)

        

        
        

len(df)

df_save = df

df[['job_title', 'job_salary', 'job_summary']]
